Crate subspace_solving

Set of modules that implement utilities for solving and verifying of solutions in Subspace Network Blockchain.

Constants

REWARD_SIGNING_CONTEXT

Signing context used for creating reward signatures by farmers.

Functions

create_chunk_signature

Create tag signature using farmer’s keypair.

create_chunk_signature_transcript

Transcript used for creation and verification of VRF signatures for chunks.

derive_global_challenge

Derive global slot challenge from global randomness.

verify_chunk_signature

Verify that chunk signature was created correctly.