Crate subspace_solving

Set of modules that implement utilities for solving and verifying of solutions in Subspace Network Blockchain.

Structs

SubspaceCodec

Subspace codec is used to encode pieces of archived history before writing them to disk and also to decode them after reading from disk.

Constants

SOLUTION_SIGNING_CONTEXT

Signing context used for creating solution signatures by farmer

Functions

create_tag

Create a commitment tag of a piece for a particular salt

derive_global_challenge

Derive global slot challenge from epoch randomness

derive_local_challenge

Derive local challenge for farmer’s public key hash from global challenge

is_commitment_valid

Check whether commitment tag of a piece is valid for a particular salt, which is used as a Proof-of-Replication