Crate subspace_solving

Set of modules that implement utilities for solving and verifying of solutions in Subspace Network Blockchain.

Structs

PieceDistance

Little-endian large integer type Distance to piece index hash from farmer identity

SubspaceCodec

Subspace codec is used to encode pieces of archived history before writing them to disk and also to decode them after reading from disk.

Enums

BatchEncodeError

CPU encoding errors

Constants

REWARD_SIGNING_CONTEXT

Signing context used for creating reward signatures by farmers.

Functions

create_tag

Create a commitment tag of a piece for a particular salt.

create_tag_signature

Create tag signature using farmer’s keypair.

create_tag_signature_transcript

Transcript used for creation and verification of VRF signatures for tags.

derive_global_challenge

Derive global slot challenge from global randomness.

derive_local_challenge

Derive local challenge for farmer from keypair and global challenge.

derive_local_challenge_and_target

Derive local challenge and target for farmer from keypair and global challenge.

derive_target

Derive challenge target from public key and local challenge.

is_tag_valid

Check whether commitment tag of a piece is valid for a particular salt, which is used as a Proof-of-Replication

verify_local_challenge

Verify local challenge for farmer’s public key that was derived from the global challenge.

verify_tag_signature

Verify that tag signature was created correctly.